ai_service/requirements.txt
fastapi
uvicorn
yt-dlp
sentence-transformers
faiss-cpu
nltk
requests